Added `parse()` to PragmaInterface; Lexer calls it
Added `parse()` to `PragmaInterface`. The lexer now calls it with the
token struct when defining the token; it is expected to return a
token struct.

A new test, `LexerTest`, validates that a pragma will be called to evaluate a
token, and is allowed to return a token, as well as that invalid tokens returned
will raise an exception.